Created a claim function for Win bets only. Had to be able to get total amounts from each pool so added a method in the race class to do that.

+++ Assignment2HorseBetting/src/inf122/horses/console/state/RacetrackState.java
@@ -36,12 +36,12 @@ public class RacetrackState
 		return rt.getRace(raceID);
 	}
 	
-	public void setTakeAmount(double takeAmount)
+	public void setTakeAmount(int takeAmount)
 	{
 		rt.setTakeAmount(takeAmount);
 	}
 	
-	public double getTakeAmount()
+	public int getTakeAmount()
 	{
 		return rt.getTakeAmount();
 	}
